requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
torch==2.5.1
torchvision==0.20.1
transformers==4.35.2
Pillow==10.1.0
numpy==1.26.2pip uninstall pydantic pydantic-core -y
pip install "pydantic>=2.7.4,<3.0.0"
pydantic>=2.7.4,<3.0.0
pandas==2.1.4